Refactor ExitStep by combining it with DestroySessionStep and flagged in the constructor
