+++ app/steps/bye/index.js
@@ -1,8 +1,8 @@
-const DestroySessionStep = require('app/core/steps/DestroySessionStep');
+const ExitStep = require('app/core/steps/ExitStep');
 
 const statusCodes = require('http-status-codes');
 
-module.exports = class ExitMarriageBroken extends DestroySessionStep {
+module.exports = class ExitMarriageBroken extends ExitStep {
   handler(req, res, next) {
     res.redirect(statusCodes.MOVED_PERMANENTLY, '/index');
     next();
