+++ app/steps/check-your-answers/index.js
@@ -9,7 +9,6 @@ const sessionBlacklistedAttributes = require('app/resources/sessionBlacklistedAt
 const ga = require('app/services/ga');
 const addressHelpers = require('../../components/AddressLookupStep/helpers/addressHelpers');
 const parseBool = require('app/core/utils/parseBool');
-const DestroySessionStep = require('app/core/steps/DestroySessionStep');
 const ExitStep = require('app/core/steps/ExitStep');
 const stepsHelper = require('app/core/helpers/steps');
 
@@ -212,7 +211,7 @@ module.exports = class CheckYourAnswers extends ValidationStep {
       delete session.nextStepUrl;
     }
 
-    if (nextStep && nextStep !== this && !(nextStep instanceof DestroySessionStep) && !(nextStep instanceof ExitStep)) {
+    if (nextStep && nextStep !== this && !(nextStep instanceof ExitStep)) {
       if (previousQuestionsRendered.length > maximumNumberOfSteps) {
         logger.errorWithReq(null, 'never_ending_loop', 'Application has entered a never ending loop. Stop attempting to build CYA template and return answers up until this point');
         return templates;
