+++ app/steps/exit/screening-questions/has-marriage-cert/index.js
@@ -1,6 +1,6 @@
-const ExitStep = require('app/core/steps/ExitStep');
+const Step = require('app/core/steps/Step');
 
-module.exports = class ExitMarriageCertificate extends ExitStep {
+module.exports = class ExitMarriageCertificate extends Step {
   get url() {
     return '/exit/screening-questions/marriage-certificate';
   }
