+++ app/steps/exit/screening-questions/has-respondent-address/index.js
@@ -1,6 +1,6 @@
-const ExitStep = require('app/core/steps/ExitStep');
+const Step = require('app/core/steps/Step');
 
-module.exports = class ExitRespondentAddress extends ExitStep {
+module.exports = class ExitRespondentAddress extends Step {
   get url() {
     return '/exit/screening-questions/respondent-address';
   }
